+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to insta and cargo-insta are documented here.
 
 - Inline snapshots only use `#` characters as delimiters when required.  #603
 
+- Experimental support for binary snapshots.  #610 (Florian Plattner)
+
 ## 1.40.0
 
 - `cargo-insta` no longer panics when running `cargo insta test --accept --workspace`